refactored logic out to make componet more simple
